fix: Fixing materialUI for nextjs
